amt to bigdecimal

+++ src/main/java/com/coding/sales/order/ItemAmtCalculatorInterface.java
@@ -2,6 +2,9 @@ package com.coding.sales.order;
 
 import com.coding.sales.beans.DiscountInforBean;
 import com.coding.sales.input.OrderItemCommand;
+
+import java.math.BigDecimal;
+
 /**
  * @author : wushengping
  * @version V1.0
@@ -11,5 +14,5 @@ import com.coding.sales.input.OrderItemCommand;
  * @date Date : 2019年07月02日 15:41
  */
 public interface ItemAmtCalculatorInterface {
-    double calculate(OrderItemCommand item, DiscountInforBean discountInforBean);
+    BigDecimal calculate(OrderItemCommand item, DiscountInforBean discountInforBean);
 }
